large file transcription through audio splitting
Whisper API has a limitation on audio file size, to overcome this limitation
we implement audio file splitting strategy using ffmpeg and its silence detection
algorithm. If the audio is longer than 45 minutes (currently hardcoded) we run
silence detection and seek for >=3s length silence then split the file in the
middle of detected silence and the call whisper on the splits. The transcribed
subtitles are then connected back.
This does only work for SRT files.